Added Interactions
Added a BP_PickupBase and BP_PedestalBase that use a BPC_Interactable, you can specify how many item you need on a pedestal and it throw an event to the game mode to check if the win conditions are met. For now I use a "pickup type" as enum to identify items. You can pick up item when you are close to them by pressing R on keyboard or A on gamepad and if you hold them you drop the item and physics start, you can pick them again if you need. Same for placing on pedestal just press the interact key but you can't remove items on it.